+++ src/lib/content/feature-activation-log.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function announceFeatureActivation(_featureKey: string, _featureName: string) {}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export function logErpDebug(_message: string, _details?: unknown) {}